Added FancyStoneSmelter Andesite/Diorite/Granite recipe
Added Andesite, Diorite, and Granite recipes to the Fancy Stone Smelter that are made with Cobblestone and Flint. I ensured to make the costs of each recipe to reflect the costs in their vanilla equivalent, and made it so that you simply can't get a larger amount of Andesite from crafting Cobblestone and Diorite than from using the factory recipe.